SKFP-16: Use Keycloak instead of Ego and convert app to typescript
