+++ .dockerignore
@@ -1,0 +1,3 @@
+Dockerfile
+.git
+node_modules
