+++ .env schema
@@ -6,5 +6,5 @@ PORT=4000
 ES_HOST=http://kf-arranger-es-qa.kids-first.io:9200
 #ES_HOST=localhost:9100
 
-# Ego host
-EGO_URL=https://ego-qa.kids-first.io
+# Keycloak host
+KEYCLOAK_URL=http://localhost:8080/auth
