+++ .eslintrc
@@ -1,24 +1,35 @@
 {
+  "env": {
+      "commonjs": true,
+      "node": true,
+      "es6": true
+  },
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+      "ecmaVersion": 2020,
+      "sourceType": "module"
+  },
   "extends": [
-    "react-app",
-    "airbnb",
-    "plugin:jsx-a11y/recommended",
-    "prettier",
-    "prettier/react"
-  ],
-  "plugins": [
-    "jsx-a11y",
-    "prettier"
+      "plugin:@typescript-eslint/recommended", 
+      "plugin:prettier/recommended"
   ],
   "rules": {
-    "semi": 0,
-    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
-    "prettier/prettier": [
-      "error", {
-        "semi": true,
-        "singleQuote": true,
-        "trailingComma": true
-      }
-    ]
+      "arrow-body-style": ["error", "as-needed"],
+      "complexity": ["warn", 16],
+      "template-curly-spacing": "off",
+      "indent": "off",
+      "default-case": 0,
+      "max-len": [
+          "error",
+          {
+              "code": 160,
+              "ignoreComments": true
+          }
+      ],
+      "no-console": ["error", { "allow": ["warn", "error"] }],
+      "no-unused-vars": "off",
+      "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }],
+      "prefer-spread": "error",
+      "prettier/prettier": ["error"]
   }
 }
