+++ .mocharc.json
@@ -1,0 +1,5 @@
+{
+    "extension": ["ts"],
+    "spec": "src/**/*.test.ts",
+    "require": "ts-node/register"
+  }
