--- index.js
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-require('dotenv').config();
-require('./babelCfg');
-require('./src/index');
