+++ package.json
@@ -4,9 +4,13 @@
   "description": "The Kids First Download Clinical Data API offers an endpoint to generate a report of clinical data based on a query built using the Kids First Portal.",
   "main": "index.js",
   "scripts": {
-    "start": "node index",
-    "watch": "nodemon index",
-    "test": "cross-env NODE_ENV=\"test\" nyc --reporter=html mocha --require ./babelCfg.js src/**/*.test.js"
+    "start": "node ./dist/index.js",
+    "watch-ts": "tsc -w",
+    "build": "tsc ",
+    "clean": "rm -rf ./dist/* ",
+    "deploy": "git push && git push --tags",
+    "lint": "eslint . --ext .ts",
+    "test": "cross-env NODE_ENV=\"test\" nyc --reporter=html mocha"
   },
   "repository": {
     "type": "git",
@@ -19,41 +23,33 @@
   },
   "homepage": "https://github.com/kids-first/kf-download-data#readme",
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.5.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-config-react-app": "^5.0.2",
-    "eslint-plugin-flowtype": "^4.3.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^2.1.2",
-    "nodemon": "^1.19.3",
-    "prettier": "^1.19.1"
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/lodash": "^4.14.171",
+    "@types/mocha": "^8.2.3",
+    "@types/node": "^16.3.2",
+    "@typescript-eslint/eslint-plugin": "^4.28.3",
+    "@typescript-eslint/parser": "^4.28.3",
+    "chai": "^4.3.4",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.29.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "mocha": "^9.0.2",
+    "nodemon": "^2.0.12",
+    "nyc": "^15.1.0",
+    "prettier": "^1.19.1",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "dependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-    "@babel/plugin-proposal-optional-chaining": "^7.6.0",
-    "@babel/preset-env": "^7.6.2",
-    "@babel/register": "^7.6.2",
     "@elastic/elasticsearch": "^7.4.0",
     "@kfarranger/middleware": "1.5.2",
-    "body-parser": "^1.19.0",
-    "chai": "^4.2.0",
-    "core-js": "^3.2.1",
     "cors": "^2.8.5",
-    "cross-env": "^6.0.3",
     "dotenv": "^8.1.0",
     "excel4node": "^1.7.2",
     "express": "^4.17.1",
-    "kfego-token-middleware": "^1.1.0",
-    "log4js": "^5.2.2",
-    "mocha": "^6.2.1",
-    "nyc": "^14.1.1",
-    "regenerator-runtime": "^0.13.3"
+    "keycloak-connect": "^14.0.0",
+    "lodash": "^4.17.21"
   }
 }
